+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.23.0-preview] - 2021-10-18
+- Fixed API Updater Configuration Validation failing due to passing a too long arguments string to the ConfigurationValidator process.
+- Changed API Validation to only run when on Windows and the editor version matches the `unity` property of the package manifest. Previous assemblies are now fetched under a new v2 prefix were all assemblies were built on Windows with an editor version matching the aforementioned `unity` property.
+
 ## [0.22.0-preview] - 2021-09-01
 - Fixed dead link to license file reference in License Validation warning message.
 - Fixed some issues with validations failing due to Windows path length limitation.